% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNumberToImage.R
\name{addNumberToImage}
\alias{addNumberToImage}
\title{Add a number to an image layer}
\usage{
addNumberToImage(
  image = NULL,
  number = NULL,
  pos_x = NULL,
  pos_y = NULL,
  number_size_factor = NULL,
  number_color = "red"
)
}
\arguments{
\item{image}{A numeric (three-dimensional) array of numbers between
0 and 1.}

\item{number}{An integer that is to be added to the image.}

\item{pos_x, pos_y}{An integer giving the x- and y-values of the starting
pixel (upper left corner).}

\item{number_size_factor}{A number for resizing the number image
to be added.}

\item{number_color}{A string reflecting the color of the number to be
added.}
}
\value{
An array (image) with added number.
}
\description{
`addNumberToImage()` adds an image of an integer to a x-y-3(rgb)-
representation of an image.
}
\examples{
# Add a red number 5 to position 500,500 of a 1024, 1024 image called test_image
addNumberToImage(test_image, 5, 500, 500)
}
