% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCilia.R
\name{detectCilia}
\alias{detectCilia}
\alias{ciliaDetect}
\alias{detectcilia}
\title{detectCilia}
\usage{
detectCilia(input_dir = NULL, cilia_color = "blue", threshold_find = 0.5,
  threshold_connect = 0.1, vicinity = NULL)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{cilia_color}{A character (color of the cilia staining)}

\item{threshold_find}{A number (minimum intensity to find cilia)}

\item{threshold_connect}{A number (minimum intensity to connect to cilia)}

\item{vicinity}{A number (neighborhood to look for pixels that belong to
a given cilium)}
}
\description{
Main function for detecting cilia in images
}
\details{
Input should be tif-format.
At first, the cilia will be located by using the max intensity method
and reconnecting them with nearby pixels with the lower threshold.
Afterwards a linear regression is being calculated for each cilium
projection which determines the length a (horizontal plane). The length b
in the vertical plane is being determined by the numbers of z-slices
where the cilium can be found. The entire length of the cilium is being
calculated with c=sqrt(a^2+b^2).
The output will be written in the current directory.
Please be aware that the coordinate system is turned by 90° to the right.
The origin of the x- and y-axes is in the upper-left corner of the image.
(The x-axis points downwards and the y-axis to the right.)
}
\examples{
\dontrun{
# Obtain all positions of cilia in every z-layer
df_cilium_information <- detectCilia(input_dir = "inst/testFiles",
                                     cilia_color = "red",
                                     threshold_find = 0.5,
                                     threshold_connect = 0.1)

# Get the length of the cilia
df_cilium_summary <- summarizeCiliaInformation(df_cilium_information,
                                               0.219647,
                                               0.20944)
}

}
\author{
Kai Budde
}
