% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectObjects.R
\name{detectObjects}
\alias{detectObjects}
\alias{objectDetect}
\alias{detectobject}
\title{detectObjects}
\usage{
detectObjects(input_dir = NULL, cilium_color = "blue", threshold = 0.1)
}
\arguments{
\item{input_dir}{A character}

\item{cilium_color}{A character}

\item{threshold}{A number}
}
\description{
Function for detecting colored objects in microscopy images
}
\details{
Input should be tif-format.
The output will be written in the current directory.
Please be aware that the coordinate system is turned by 90° to the right.
The origin of the x- and y-axes is in the upper-left corner of the image.
(The x-axis points downwards and the y-axis to the right.)
}
\examples{
\dontrun{
detectObject(input.file = "example.tif")
}


}
\author{
Kai Budde
}
