% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editImage.R
\name{editImage}
\alias{editImage}
\title{Calculate binary mask}
\usage{
editImage(image = NULL, object_color = NULL, threshold = NULL)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1.}

\item{object_color}{A character being the color of the staining of the
object to be masked.}

\item{threshold}{A number that determines the brightness of a pixel to
be counted as a cilium pixel.}
}
\value{
A binary image.
}
\description{
`editImage()` masks an image layer depending on a threshold.
}
