% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLayer.R
\name{getLayer}
\alias{getLayer}
\title{Get a specific layer of the image}
\usage{
getLayer(image = NULL, layer = NULL)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1.}

\item{layer}{A character being the color of the layer.}
}
\value{
An array (specific channel of image).
}
\description{
`getLayer()` extracts a specific layer of an image (being x-y-3(rgb)-
representation).
}
