% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resizeImage.R
\name{resizeImage}
\alias{resizeImage}
\title{Resizes an image}
\usage{
resizeImage(image = NULL, number_size_factor = 1)
}
\arguments{
\item{image}{A one to three-dimensional array of numbers between 0 and 1}

\item{number_size_factor}{A number being the factor for resizing the
image of a number.}
}
\value{
An array (resized image).
}
\description{
`resizeImage()` eesizes an image with a given scaling factor
}
