% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCiliaInformation.R
\name{summarizeCiliaInformation}
\alias{summarizeCiliaInformation}
\title{Summarize the information of found cilia}
\usage{
summarizeCiliaInformation(
  df_cilium_information = df_cilium_information,
  min_cilium_area = min_cilium_area,
  pixel_size = pixel_size,
  slice_distance = slice_distance
)
}
\arguments{
\item{df_cilium_information}{A data frame with found cilia.}

\item{min_cilium_area}{A number for the minimum number of pixels needed
to represent a cilium.}

\item{pixel_size}{A number depicting the size of one pixel in micrometers.}

\item{slice_distance}{A number depicting the distance of two consecutive
z-stack layers in micrometers.}
}
\value{
A tibble with summarized cilium information (lengths).
}
\description{
`summarizeCiliaInformation()` summarizes the information of cilia found
by the main function detectCilia.R. The output is a data frame with the
lengths of the found cilia.
}
