% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCilia.R
\name{detectCilia}
\alias{detectCilia}
\alias{ciliaDetect}
\alias{detectcilia}
\title{detectCilia}
\usage{
detectCilia(
  input_dir = NULL,
  cilia_color = "red",
  threshold_find = 0.9,
  threshold_connect = 0.5,
  vicinity = NULL,
  min_size = 3,
  max_size = 100
)
}
\arguments{
\item{input_dir}{A character (directory that contains all images)}

\item{cilia_color}{A character (color of the cilia staining)}

\item{threshold_find}{A number (minimum intensity to find cilia)}

\item{threshold_connect}{A number (minimum intensity to connect to cilia)}

\item{vicinity}{A number (neighborhood to look for pixels that belong to
a given cilium)}

\item{min_size}{A number (gives the minimum size of a cilium to be
detected)}

\item{max_size}{A number (gives the maximum size of a cilium to be
detected)}
}
\description{
Main function for detecting cilia in images
}
\details{
Input should be tif-format.
At first, the cilia will be located by using the max intensity method
and reconnecting them with nearby pixels with the lower threshold.
Afterwards a linear regression is being calculated for each cilium
projection which determines the length a (horizontal plane). The length b
in the vertical plane is being determined by the numbers of z-slices
where the cilium can be found. The entire length of the cilium is being
calculated with c=sqrt(a^2+b^2).
The output will be written in the current directory.
Please be aware that the coordinate system is turned by 90° to the right.
The origin of the x- and y-axes is in the upper-left corner of the image.
(The x-axis points downwards and the y-axis to the right.)
}
\author{
Kai Budde
}
