% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectCilia.R
\name{detectCilia}
\alias{detectCilia}
\alias{ciliaDetect}
\alias{detectcilia}
\title{detectCilia}
\usage{
detectCilia(
  input_dir_tif = NULL,
  input_file_czi = NULL,
  cilium_color = "red",
  nucleus_color = "blue",
  projection_method = "mean",
  threshold_by_density_of_cilium_pixels = TRUE,
  threshold_find = NULL,
  threshold_connect = NULL,
  vicinity_combine = NULL,
  vicinity_connect = NULL,
  min_cilium_area = NULL,
  max_cilium_area = NULL,
  number_size_factor = NULL,
  pixel_size = NULL,
  slice_distance = NULL,
  nuc_mask_width_heigth = NULL
)
}
\arguments{
\item{input_dir_tif}{A character (directory that contains z-stack sclices
in the tif format)}

\item{input_file_czi}{A character (a file with the z-stack image as czi)}

\item{cilium_color}{A character (color of the cilium staining)}

\item{nucleus_color}{A character (color of the nuclei staining)}

\item{projection_method}{A character (defines the method for the z projection:
either "mean" or "max")}

\item{threshold_by_density_of_cilium_pixels}{A Boolean (disregard the
threshold values if true and instead use a custom function to calculate
the thresholds by looking at the density of cilium color pixels found in
the image)}

\item{threshold_find}{A number (minimum intensity to find cilia)}

\item{threshold_connect}{A number (minimum intensity to connect to
already detected cilium)}

\item{vicinity_combine}{A number (neighborhood to look for strucutres
belonging to the same cilium (exclusion volume))}

\item{vicinity_connect}{A number (neighborhood to look for pixels that
belong to a given cilium (inclusion volume))}

\item{min_cilium_area}{A number (gives the minimum area of a cilium to be
detected)}

\item{max_cilium_area}{A number (gives the maximum area of a cilium to be
detected)}

\item{number_size_factor}{A number (factor for resizing the digit image)}

\item{pixel_size}{A number (size of one pixel in micrometer)}

\item{slice_distance}{A number (distance of two consecutive slices in
z-direction in micrometer)}

\item{nuc_mask_width_heigth}{A number (width and heigth for detecting
nuclei)}
}
\description{
Main function for detecting cilia in images
}
\details{
Input should be tif-format.
At first, the cilia will be located by using the max intensity method
and reconnecting them with nearby pixels with the lower threshold.
Afterwards a linear regression is being calculated for each cilium
projection which determines the length a (horizontal plane). The length b
in the vertical plane is being determined by the numbers of z-slices
where the cilium can be found. The entire length of the cilium is being
calculated with c=sqrt(a^2+b^2).
The output will be written in the current directory.
Please be aware that the coordinate system is turned by 90° to the right.
The origin of the x- and y-axes is in the upper-left corner of the image.
(The x-axis points downwards and the y-axis to the right.)
}
\author{
Kai Budde
}
