% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editImage.R
\name{editImage}
\alias{editImage}
\alias{editimage}
\alias{imageedit}
\alias{imageEdit}
\title{editImage}
\usage{
editImage(image = NULL, cilia_color = NULL, threshold = NULL)
}
\arguments{
\item{image}{An three-dimensional array of numbers between 0 and 1}

\item{cilia_color}{A character (color of the staining of the cilia)}

\item{threshold}{A character (that determine tha brightness of a pixel
to be counted)}
}
\description{
Get the layer of the primary cilia
}
\details{
By using an x-y-3(rgb)-representation of an image, you can
extract the information of the cilia in an image.
}
\author{
Kai Budde
}
