% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeCiliaInformation.R
\name{summarizeCiliaInformation}
\alias{summarizeCiliaInformation}
\alias{summarizeCiliumInformation}
\alias{summarizeciliainformation}
\title{summarizeCiliaInformation}
\usage{
summarizeCiliaInformation(
  df_cilium_information = df_cilium_information,
  min_cilium_area = min_cilium_area,
  pixel_size = pixel_size,
  slice_distance = slice_distance
)
}
\arguments{
\item{df_cilium_information}{A data frame}

\item{min_cilium_area}{A number (min number of pixels needed for a layer being part of a cilium)}

\item{pixel_size}{A number (size of one pixel in micrometer)}

\item{slice_distance}{A number (distance of two consecutive slices in
z-direction in micrometer)}
}
\value{
A tibble with summarized cilium information (lengths).
}
\description{
Summarizes the information of found cilia in images
}
\details{
Input should be the result of detectCilia.R.
The output is a data frame with the lengths of the found cilia.
}
\author{
Kai Budde-Sagert
}
